% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric_set.R
\name{get_metric_set}
\alias{get_metric_set}
\title{Get metrics for multiple h2o models on single training or testing frame}
\usage{
get_metric_set(df, model_list)
}
\arguments{
\item{df}{a training or testing frame. Note, this is not an h2o frame.}

\item{model_list}{a list of h2o model objects}
}
\value{
a dataframe of metrics for each model in model_list
}
\description{
Get metrics for multiple h2o models on single training or testing frame
}
\examples{

library(tidyerse)
library(h2o)
h2o.init()
iris_h2o <- as.h2o(iris)
gbm_h2o <- h2o.gbm(y = "Sepal.Length", training_frame = iris_h2o)
glm_h2o <- h2o.glm(y = "Sepal.Length", training_frame = iris_h2o)
metrics <- get_h2o_metric_set(iris, list(gmb_h2o, glm_h2o))

}
