% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get h2o model performance metrics}
\usage{
get_metrics(df, h2o_model)
}
\arguments{
\item{df}{a training or testing data frame or tibble. Note, this is not an h2o frame.}

\item{h2o_model}{the h2o model object}
}
\value{
a data frame of model metrics for input data frame
}
\description{
Get h2o model performance metrics
}
\examples{

library(tidyerse)
library(h2o)
h2o.init()
iris_h2o <- as.h2o(iris)
gbm_h2o <- h2o.gbm(y = "Sepal.Length", training_frame = iris_h2o)
metrics <- get_metrics(iris, gbm_h2o)

}
