% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_set.R
\name{plot_metric_set}
\alias{plot_metric_set}
\title{Plot metric set for list of h2o models}
\usage{
plot_metric_set(df, model_list)
}
\arguments{
\item{df}{a training or testing frame. Note this is not an h2o frame.}

\item{model_list}{list of h2o model objects}
}
\value{
ggplot object (scatterplot) show all models across all applicable metrics
}
\description{
Plot metric set for list of h2o models
}
\examples{

library(tidyerse)
library(h2o)
h2o.init()
iris_h2o <- as.h2o(iris)
gbm_h2o <- h2o.gbm(y = "Sepal.Length", training_frame = iris_h2o)
glm_h2o <- h2o.glm(y = "Sepal.Length", training_frame = iris_h2o)
metric_plot <- plot_metric_set(iris, list(gbm_h2o, glm_h2o))
metric_plot

}
